Further cleanup of TracingViewModel.kt, TracingRepository.kt Still not quite there yet, too many calls going into every direction, back and forth, instead of just top->down, but we are getting there. We want data flow from business logic code to the UI, keep the UI KISS.

+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/ui/tracing/card/TracingCardStateTest.kt
@@ -37,7 +37,7 @@ class TracingCardStateTest : BaseTest() {
         riskLevelLastSuccessfulCalculation: Int = 0,
         matchedKeyCount: Int = 0,
         daysSinceLastExposure: Int = 0,
-        activeTracingDaysInRetentionPeriod: Int = 0,
+        activeTracingDaysInRetentionPeriod: Long = 0,
         lastTimeDiagnosisKeysFetched: Date? = mockk(),
         isBackgroundJobEnabled: Boolean = false,
         isManualKeyRetrievalEnabled: Boolean = false,
